Move the Inserter into its own component for readability
