+++ packages/edit-site/src/components/secondary-sidebar/inserter-sidebar.js
@@ -1,0 +1,38 @@
+/**
+ * WordPress dependencies
+ */
+import { useDispatch } from '@wordpress/data';
+import { Button } from '@wordpress/components';
+import { __experimentalLibrary as Library } from '@wordpress/block-editor';
+import { close } from '@wordpress/icons';
+import {
+	useViewportMatch,
+	__experimentalUseDialog as useDialog,
+} from '@wordpress/compose';
+
+export default function InserterSidebar() {
+	const { setIsInserterOpened } = useDispatch( 'core/edit-site' );
+
+	const isMobile = useViewportMatch( 'medium', '<' );
+	const [ inserterDialogRef, inserterDialogProps ] = useDialog( {
+		onClose: () => setIsInserterOpened( false ),
+	} );
+
+	return (
+		<div
+			ref={ inserterDialogRef }
+			{ ...inserterDialogProps }
+			className="edit-site-editor__inserter-panel"
+		>
+			<div className="edit-site-editor__inserter-panel-header">
+				<Button
+					icon={ close }
+					onClick={ () => setIsInserterOpened( false ) }
+				/>
+			</div>
+			<div className="edit-site-editor__inserter-panel-content">
+				<Library showInserterHelpPanel shouldFocusBlock={ isMobile } />
+			</div>
+		</div>
+	);
+}
